+++ aws/rds/db_password.tf
@@ -1,21 +1,21 @@
 resource "random_password" "this" {
   count = var.use_secret_manager ? 0 : 1
 
-  length  = 24
+  length  = var.password_length
   special = false
 }
 
 data "aws_secretsmanager_random_password" "this" {
   count = var.use_secret_manager ? 1 : 0
 
-  password_length     = 32
+  password_length     = var.password_length
   exclude_punctuation = true
 }
 
 resource "aws_secretsmanager_secret" "this" {
   count = var.use_secret_manager ? 1 : 0
 
-  name = local.db_password_name
+  name = var.secret_manager_db_password_name
 }
 
 resource "aws_secretsmanager_secret_version" "this" {